@@ -14,16 +14,18 @@ description: Hackernews Flutter Application
 version: 1.1.1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
+  flutter: ">=3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   http: 0.12.2
   logger: ^0.8.2
-  webview_flutter: ^0.3.19+4
-  url_launcher: ^5.4.1
-  share: ^0.6.3+5
+  path_provider: ^2.0.6
+  webview_flutter: ^4.0.5
+  webview_flutter_android: ^3.0.0
+  webview_flutter_wkwebview: ^3.0.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
