@@ -1,4 +1,4 @@
-package `in`.codesmell.hacker_news
+package `in`.codesmell.hacker_news_tv
 
 import io.flutter.embedding.android.FlutterActivity
 
